changed: move Eval and restrictEval to WellInterfaceIndices
